opencl_vendor_mesa now excludes all OpenGL context sharing functions
Currently Mesa does not support context sharing with OpenGL.
The related functions are missing from the Mesa library and will
cause linker errors if referenced.

Constants and structs related to the context sharing with OpenGL
still included. Reasoning: Easier to develop libraries / applications
that in some configurations require OpenGL context sharing. With
the constants and structs included, there is potentially less
functionality that the library / application needs to exclude
with Mesa.